Move Transmission to libvirt/Wireguard
There were 2 main problems with the previous setup using Docker and OpenVPN:
- Stability: The OpenVPN tunnel wasn't stable anymore since more than 1 month. Restarting the container temporary resolved the issue. Unfortunately, no clear logs.
- Ensure traffic only uses VPN: I didn't like the "thin" separation of VPN and normal traffic.

The new setup uses libvirt and Wireguard to combat both issues.

Since Transmission is only accessible to the host (by design), I added an internal proxy to Traefik to route traffic over the host. This can be used as a quick and dirty jumphost from other Docker containers.