Kubo: Make Mosquitto persistent
When HA/Zigbee2MQTT/Mosquitto is restarted, all Zigbee devices and their are "forgotten" on HA. Zigbee2MQTT republishes the devices and states on start of HA, but somehow it's not working. So, I expect the persistence on Mosquitto is required for it to work. Let's see if it solves the issue.